+++ core-ng/src/main/java/core/framework/impl/web/request/PathParams.java
@@ -15,11 +15,11 @@ public final class PathParams {
     private final Map<String, String> params = Maps.newHashMap();
 
     public void put(String name, String value) {
-        if (value.length() == 0) throw new BadRequestException(format("path param must not be empty, name={}, value={}", name, value));
+        if (value.length() == 0) throw new BadRequestException(format("path param must not be empty, name={}, value={}", name, value), "INVALID_HTTP_REQUEST");
         try {
             params.put(name, Encodings.decodeURIComponent(value));  // value here is not decoded, see io.undertow.UndertowOptions.DECODE_URL and core.framework.impl.web.HTTPServer
         } catch (IllegalArgumentException e) {
-            throw new BadRequestException(e.getMessage(), BadRequestException.DEFAULT_ERROR_CODE, e);
+            throw new BadRequestException(e.getMessage(), "INVALID_HTTP_REQUEST", e);
         }
     }