+++ core-ng/src/main/java/core/framework/web/exception/NotFoundException.java
@@ -10,24 +10,17 @@ import core.framework.log.Severity;
  */
 @ResponseStatus(HTTPStatus.NOT_FOUND)
 public final class NotFoundException extends RuntimeException implements ErrorCode {
-    public static final String DEFAULT_ERROR_CODE = "NOT_FOUND";
-
     private static final long serialVersionUID = 8663360723004690205L;
 
     private final String errorCode;
 
     public NotFoundException(String message) {
         super(message);
-        errorCode = DEFAULT_ERROR_CODE;
+        errorCode = "NOT_FOUND";
     }
 
     public NotFoundException(String message, String errorCode) {
         super(message);
-        this.errorCode = errorCode;
-    }
-
-    public NotFoundException(String message, String errorCode, Throwable cause) {
-        super(message, cause);
         this.errorCode = errorCode;
     }