+++ core-ng/src/main/java/core/framework/web/websocket/Channel.java
@@ -15,7 +15,7 @@ public interface Channel {
     void leave(String room);
 
     interface Context {
-        Object get(String key);
+        Object get(String key); // channel is stateful, context key usually be static put onConnect, so in most of case get(key) expect result, that's why here is designed to return Object, not Optional<T>
 
         void put(String key, Object value);
     }